+++ graphrag/callbacks/console_workflow_callbacks.py
