+++ graphrag/index/run/workflow.py
@@ -15,12 +15,13 @@ from datashaper import (
     WorkflowCallbacksManager,
 )
 
+from graphrag.callbacks.progress_workflow_callbacks import ProgressWorkflowCallbacks
 from graphrag.index.context import PipelineRunContext
 from graphrag.index.emit.table_emitter import TableEmitter
 from graphrag.index.run.profiling import _write_workflow_stats
 from graphrag.index.storage.typing import PipelineStorage
 from graphrag.index.typing import PipelineRunResult
-from graphrag.logging import ProgressLogger, ProgressWorkflowCallbacks
+from graphrag.logging import ProgressLogger
 from graphrag.utils.storage import _load_table_from_storage
 
 log = logging.getLogger(__name__)
