+++ graphrag/query/llm/base.py
@@ -7,16 +7,7 @@ from abc import ABC, abstractmethod
 from collections.abc import AsyncGenerator, Generator
 from typing import Any
 
-
-class BaseLLMCallback:
-    """Base class for LLM callbacks."""
-
-    def __init__(self):
-        self.response = []
-
-    def on_llm_new_token(self, token: str):
-        """Handle when a new token is generated."""
-        self.response.append(token)
+from graphrag.callbacks.llm_callbacks import BaseLLMCallback
 
 
 class BaseLLM(ABC):
